requirements.txt
flask[async]
flask-cors
flask_jwt_extended
python-dotenv
pymongo
aiofiles
bcrypt
pydantic
pandas
celery
starlette
uvicorn
redis